+++ src/EventSubscriber/PersonEventSubscriber.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace Dbp\Relay\BasePersonConnectorLdapBundle\EventSubscriber;
 
 use Dbp\Relay\BasePersonConnectorLdapBundle\Event\PersonPostEvent;
+use Dbp\Relay\BasePersonConnectorLdapBundle\Event\PersonPreEvent;
 use Dbp\Relay\CoreBundle\LocalData\AbstractLocalDataEventSubscriber;
 
 class PersonEventSubscriber extends AbstractLocalDataEventSubscriber
@@ -12,6 +13,7 @@ class PersonEventSubscriber extends AbstractLocalDataEventSubscriber
     public static function getSubscribedEventNames(): array
     {
         return [
+            PersonPreEvent::class,
             PersonPostEvent::class,
         ];
     }